requirements.txt
streamlit
Pillow
langchain-google-genai
langchain-core
google-generativeai
python-dotenv